All problems sovled

+++ src/Distance/Distance.cpp
@@ -20,7 +20,7 @@ template <typename T>
 double Euclidean_Distance(const T & v1,const T &v2){
  	long double sum = 0;
  	for(unsigned int i = 0;i<v1.size();i++){
- 		sum += pow((v1[i] - v2[i]),2);
+ 		sum += pow((double)(v1[i] - v2[i]),2);
  	}
  	return sqrt(sum);
  };
